add new features from master

+++ MyCinema_App/src/main/java/app/controller/AdminRoomsController.java
@@ -8,10 +8,20 @@ import app.database.service.IRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.URL;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.nio.file.StandardCopyOption;
 import java.util.List;
 import java.util.Optional;
 
@@ -37,18 +47,42 @@ public class AdminRoomsController
     public String deleteRoom(@RequestParam(name = "room-ids", required = false) List<String> roomIds)
     {
         if (roomIds != null)
+        {
             for (String id : roomIds)
-                repository.deleteById(id);
+            {
+                Optional<CinemaRoom> optionalRoom = repository.findById(id);
+
+                if (optionalRoom.isPresent())
+                {
+                    repository.deleteById(id);
+
+                    try
+                    {
+                        Files.deleteIfExists(Paths.get("src/main/resources/static/images/cinema-rooms/" + optionalRoom.get().getId() + ".jpg"));
+                    }
+                    catch (IOException e)
+                    {
+                        System.err.println("Couldn't delete file: " + e);
+                    }
+                }
+            }
+        }
 
         return "redirect:/admin-rooms";
     }
 
     @PostMapping(value = "/admin-add-room")
     public String addRoom(@RequestParam(name = "room-name", required = true) String roomName,
-                          @RequestParam(name = "room-image-path", required = true) String roomImagePath)
+                          @RequestParam(name = "room-image", required = true) MultipartFile file)
     {
-        if (!roomName.trim().isEmpty() && !roomImagePath.trim().isEmpty())
-            repository.save(new CinemaRoom(roomName, roomImagePath));
+        if (roomName.trim().isEmpty())
+            return "redirect:/admin-rooms";
+
+        CinemaRoom cinemaRoom = new CinemaRoom(roomName);
+        repository.save(cinemaRoom);
+
+        if (!saveImage(file, "src/main/resources/static/images/cinema-rooms", cinemaRoom.getId() + ".jpg"))
+            repository.delete(cinemaRoom);
 
         return "redirect:/admin-rooms";
     }
@@ -69,7 +103,7 @@ public class AdminRoomsController
 
     @PostMapping(value = "/admin-rooms-submit-edit")
     public String saveRoom(@RequestParam(name = "new-room-name") String newRoomName,
-                           @RequestParam(name = "new-room-image-path") String newRoomImagePath,
+                           @RequestParam(name = "new-room-image") MultipartFile file,
                            @RequestParam(name = "room-id") String roomId)
     {
         Optional<CinemaRoom> optionalRoom = repository.findById(roomId);
@@ -81,12 +115,56 @@ public class AdminRoomsController
             if (!newRoomName.trim().isEmpty())
                 room.setName(newRoomName);
 
-            if (!newRoomImagePath.trim().isEmpty())
-                room.setImagePath(newRoomImagePath);
+            if (!file.isEmpty())
+                saveImage(file, "src/main/resources/static/images/cinema-rooms", roomId + ".jpg");
 
             repository.save(room);
         }
 
         return "redirect:/admin-rooms";
+    }
+
+    @RequestMapping(value = "/images/cinema-rooms/{imageId}")
+    @ResponseBody
+    public byte[] getImage(@PathVariable String imageId, HttpServletRequest request)
+    {
+        Path path = Paths.get("src/main/resources/static/images/cinema-rooms/" + imageId);
+
+        try {
+            byte[] data = Files.readAllBytes(path);
+            return data;
+        }
+        catch (IOException e)
+        {
+            System.out.println(e);
+        }
+
+        return null;
+    }
+
+    private boolean saveImage(MultipartFile inputFile, String saveLocation, String saveFilename)
+    {
+        String inputFilename = StringUtils.cleanPath(inputFile.getOriginalFilename());
+        Path path = Paths.get(saveLocation);
+
+        try
+        {
+            if (inputFile.isEmpty())
+                throw new IOException("Failed to store empty file " + inputFilename);
+            else if (inputFilename.contains(".."))
+                throw new IOException("Cannot store file with relative path outside current directory " + inputFilename);
+
+            try (InputStream inputStream = inputFile.getInputStream())
+            {
+                Files.copy(inputStream, path.resolve(saveFilename), StandardCopyOption.REPLACE_EXISTING);
+            }
+        }
+        catch (IOException e)
+        {
+            System.err.println("Failed to store file " + inputFilename + ". Reason:\n" + e);
+            return false;
+        }
+
+        return true;
     }
 }
